Moved some files and added styles at the info on HotelPage
