+++ Booking-front/src/css/index.scss~
@@ -919,7 +919,7 @@
   .hotel-page {
     width: 100%;
     padding-top: 45px;
-    padding-bottom: 92px;
+    margin-bottom: 92px;
 
     .hotel-section {
       //border: 1px solid #000;
