+++ Booking-front/src/interfaces/hotel/IAddressCreate.ts
@@ -1,0 +1,8 @@
+
+export interface IAddressCreate {
+    street: string;
+    houseNumber: string;
+    floor?: string;
+    apartmentNumber?: string;
+    cityId: number;
+}
