+++ Booking-front/src/interfaces/hotel/IHotel.ts
@@ -1,5 +1,5 @@
 import { HotelCategory } from "interfaces/hotelCategories";
-import { IAddress, IAddressCreate } from "interfaces/hotel/IAddress.ts";
+import { IAddress } from "interfaces/hotel/IAddress.ts";
 import IHotelAmenity from "interfaces/hotelAmenity/IHotelAmenity.ts";
 import IPhoto from "interfaces/hotel/IPhoto.ts";
 import IRealtorShortInfo from "interfaces/hotel/IRealtorShortInfo.ts";
@@ -22,20 +22,4 @@ export interface IHotel {
     realtor: IRealtorShortInfo;
     hotelAmenities: IHotelAmenity[];
     photos: IPhoto[];
-}
-
-export interface IHotelCreate {
-    name: string;
-    description: string;
-    arrivalTimeUtcFrom: string;
-    arrivalTimeUtcTo: string;
-    departureTimeUtcFrom: string;
-    departureTimeUtcTo: string;
-    isArchived?: boolean;
-    address: IAddressCreate;
-    categoryId: number;
-    hotelAmenityIds?: number[];
-    breakfastIds?: number[];
-    staffLanguageIds?: number[];
-    photos: File[];
 }
