+++ Booking-front/src/interfaces/hotel/IHotelCreate.ts
@@ -1,0 +1,17 @@
+import { IAddressCreate } from "interfaces/hotel/IAddressCreate.ts";
+
+export interface IHotelCreate {
+    name: string;
+    description: string;
+    arrivalTimeUtcFrom: string;
+    arrivalTimeUtcTo: string;
+    departureTimeUtcFrom: string;
+    departureTimeUtcTo: string;
+    isArchived?: boolean;
+    address: IAddressCreate;
+    categoryId: number;
+    hotelAmenityIds?: number[];
+    breakfastIds?: number[];
+    staffLanguageIds?: number[];
+    photos: File[];
+}
